Comentando cod

+++ include/tAlgoritmosOrdenacao.h
@@ -5,14 +5,14 @@
 
 #define TRUE 1
 
-void selectionsort(int *vet, int tam, int T, int *comp, int *trocas); //
+void selectionsort(int *vet, int tam, int T, long long int *comp, long long int *trocas); //
 
-void insertionsort(int *vet, int tam, int T, int *comp, int *trocas); 
+void insertionsort(int *vet, int tam, int T, long long int *comp, long long int *trocas); 
 
-void shellSort(int *vet, int tam, int t,int *comp, int *trocas);
+void shellSort(int *vet, int tam, int t,long long int *comp, long long int *trocas);
 
-void quickSort(int *vet, int esq, int dir, int *comp, int *trocas, int T); //
+void quickSort(int *vet, int esq, int dir, long long int *comp, long long int *trocas, int T); //
 
-void heapSort(int *vet, int n, int T , int *comp,int *trocas);
+void heapSort(int *vet, int n, int T , long long int *comp,long long int *trocas);
 
 #endif
